+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Version: 1.7.43
 
 ### New
+ - Fixed done_cells in merkle update
+ - Fixed collect_used_paths_cells